evershine/requirements.txt
asgiref==3.2.10
Django==3.0.8
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
passlib==1.7.2
PyJWT==1.7.1
PyMySQL==0.10.0
pytz==2020.1
sqlparse==0.3.1
